Separate out mocks file
